+++ src/main/java/org/apache/sysds/runtime/ooc/cache/legacy/DeferredReadQueue.java
@@ -17,7 +17,10 @@
  * under the License.
  */
 
-package org.apache.sysds.runtime.ooc.cache;
+package org.apache.sysds.runtime.ooc.cache.legacy;
+
+import org.apache.sysds.runtime.ooc.cache.BlockEntry;
+import org.apache.sysds.runtime.ooc.cache.BlockKey;
 
 import java.util.ArrayList;
 import java.util.Collections;
