+++ src/main/java/org/apache/sysds/runtime/ooc/memory/CachedAllowance.java
@@ -25,7 +25,7 @@ import org.apache.sysds.runtime.instructions.ooc.OOCStream;
 import org.apache.sysds.runtime.instructions.spark.data.IndexedMatrixValue;
 import org.apache.sysds.runtime.ooc.cache.BlockKey;
 import org.apache.sysds.runtime.ooc.cache.OOCCacheManager;
-import org.apache.sysds.runtime.ooc.cache.OOCCacheScheduler;
+import org.apache.sysds.runtime.ooc.cache.legacy.OOCCacheScheduler;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
