+++ src/test/java/org/apache/sysds/test/component/ooc/cache/SourceBackedCacheSchedulerTest.java
@@ -29,7 +29,7 @@ import org.apache.sysds.runtime.ooc.cache.BlockEntry;
 import org.apache.sysds.runtime.ooc.cache.BlockKey;
 import org.apache.sysds.runtime.ooc.cache.BlockState;
 import org.apache.sysds.runtime.ooc.cache.OOCIOHandler;
-import org.apache.sysds.runtime.ooc.cache.OOCLRUCacheScheduler;
+import org.apache.sysds.runtime.ooc.cache.legacy.OOCLRUCacheScheduler;
 import org.apache.sysds.runtime.ooc.cache.OOCMatrixIOHandler;
 import org.apache.sysds.test.AutomatedTestBase;
 import org.apache.sysds.test.TestConfiguration;
